Fixed Catalogs and Catalog Items tree building
Ensured that all 3 trees in the Catalog Explorer, prior to the actual
tree building set these values - x_active_tree and x_active_accord for
the respective trees.
Setting these values is important for each tree as it ensures proper
tree building.
This was missing previously causing a crash when a user with permissions
 to only the 'Catalogs' accordion logged in.

https://bugzilla.redhat.com/show_bug.cgi?id=1130301