Ensured that dynatree cookies get deleted after a session reset.
The session reset that happens after a LDAP user switches groups, clears
out all the dynatree sandboxes. However the cookies are left intact
which cause an issue with the dynatrees, making them lose some
functionality. For e.g. the trees expand the nodes, but nothing
happens when the user clicks on them.

https://bugzilla.redhat.com/show_bug.cgi?id=1102250
https://bugzilla.redhat.com/show_bug.cgi?id=1086468