Last value in chart zeroed by hard
Right now, the last value in the chart was always made zero.

The correct condition is: if there is at least one not nil value,
it can't make the last value zero. Otherwise all values are nil
and it will change the last value to zero(due to chart cannot
handle all values to be nil).