remove "cancel" handling from `create`
This patch allows us to avoid doing a roundtrip to the server just to
gereate a string and redirect the user.  It implements the "cancel"
button just with javascript.  This means that clicking "cancel" will
simply take the user to the page they want in 1 request rather than 2
requests (plus RJS).