Topology widget for containers provider.
Displays a graph of container related entities and
their relationships, as well as cross provider relationships
with vms and hosts if such exist.
The patch adds a new UI screen for the graph with legend and refresh
button, menu, navigation and backend logic (RoR and Angular controllers)
that enables the topology.
[Original commit d7415ef5d2acffc2258c540936be26fe7550d7b4]