Unable to add Infrastructure and cloud providers
Added a permission_prefix method to ems_common that returns the controller_name appropriately for ems_infra, ems_cloud & ems_container controllers. This method is used together with the action as the feature id in asserting privileges:  assert_privileges("#{permission_prefix}_edit"). The permission_prefix method was added to application_controller but can be overridden in each controller class as needed in the future.
Adding Rspecs.

https://bugzilla.redhat.com/show_bug.cgi?id=1238530

Conflicts:
	app/controllers/ems_common.rb