configure the application with a permission store and use it
this commit comfigures the application with a permission store.  It's
set to a null store that always returns true in test.  In other
environments, it reads from `vmdb/config/permissions.yml`.

This will probably make the website nearly useless because the yaml file
doesn't contain all the permissions that the application actually uses.
[Original commit 85406c43244230e9fc62385ec6f2c2f4c91cb9cf]