Availability Zone Block Storage Disk Usage and Capacity
Adds two fields to the availability zone detail page showing
disk usage and capacity for OpenStack clouds.

Note disk usage is calculated by summing up all of the hosts
that are part of the BlockStorage cluster. It does not take
into account the availability zone of each Cinder storage
node. If multiple Cinder availability zones are deployed, the
disk capacity numbers will not be accurate. But at the moment,
RHELOSP deployments only supports a single Cinder availability
zone so this is fine for now.

When multiple Cinder availability zones are supported, the
refresh process will need to database each Cinder node's
availability zone. And the disk capacity calculation will
need to be adjusted to take zones into account.