Trimmed leading space from the "Assign To" pull down values
With patternfly select conversions, leading space in front of "The Enterprise" was causing automate tests to fail. Leading space was added to this entry back when ruby hashes did not retain the order to keep this entry at the top of the pull down. Removed sort from view so values can be displayed in the order they were added in the initial Hash.

https://bugzilla.redhat.com/show_bug.cgi?id=1273654