Speed up tree_builder_ops_rbac_spec by eliminating extra seeds.
Previously, the test seeded all product features and roles, which is not
necessary.  Instead, we seed only the features we need.  Additionally,
the test worked against specific users, but instead we rely on any user
with a particular role, which is more important in these use cases.

This set of specs were SQL intensive with 18639 SQL statements. The
timings for this drop from 15.97s to 1.5s