Speed up rbac_tree_spec by removing seeding of product features.
We still will seed a small subset of the tree, but not the whole thing
as it's overkill.  In addition, remove the unnecessary controller
testing and focus this unit test on the OpsController::RbacTree class
itself.

This spec was a SQL intensive spec with 4360 SQL statements.  This drops
the timing of the spec from 3.44s to 0.64s