Fix cluster level block and object storage views
A undefined method X_disk_usage error occurs when a cloud
provider is not linked to an infrastructure provider.

This patch checks that a cloud provider is present, if it
is not, then disk usage is not displayed for the cluster.

Fixes BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1279435
[Original commit 02cdb573bb8601aa3f67e018834d699881b66994]