Remove unused settings from hostdefaults.
The UI for using the 3 hostdefaults settings was removed in 2010 in
c2fe77bd3d6bd62e059c9255267847d52e5b4497 (and even then, the code was
wrapped in `if false` anyway). The controllers were subsequently cleaned
up except for one or two minor lines in 2015.

Thus, these settings have not been used in all that time.  All of the
features they represent are handled by other means.

So, we can remove the settings values for all 3 as well as the
corresponding accessors. The only callers of those accessors only ran
meaningful code if those values had changed, but since there was no way to
change those values, all of those methods are noops.  Thus, we can remove
all of the callers.  Also, there was a schedule running one of those
methods, which can also be removed, along with its configuration value
from the vmdb.tmpl.yml.