Fix availability zone views when infra provider is not attached
When an OpenStack infrastructure provider is not attached to
the cloud provider, an "undefined method 'infra_ems'" error is
thrown. This fix adds a nil check on provider. If it is nil,
the disk capacity is not shown for the availability zone.

Fixes BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1291774