angular directives - split start_form_div from auto-focus
in Host form editor, when more than one item is selected, the div that has auto-focus is not actually rendered (ruby-side), so the whole div was never un-hidden

this splits the start-form-div functionality into a separate directive, and moves it to the form in that case.

Other uses of start-form-div can stay the same.

https://bugzilla.redhat.com/show_bug.cgi?id=1296258