FIX: Approve Vm provisioning in UI
https://bugzilla.redhat.com/show_bug.cgi?id=1301631

The workflow requester is set to the current_user.
The group context for the request was then being set in the the
current_user, corrupting it.

Changes:

1. MiqRequestWorkflow#requester = MiqRequest#requester
so the correct dialogs will be displayed.
2. Requester is cloned before setting the group as a precaution.

This assumes the viewer and administrator can see the resources that
the requester has access to view as well.