dialog_field_refresh - don't construct new DOM elements textually, use jQuery
this fixes `refreshRadioList` and `addOptionsToDropDownList` to always use jQuery to create the new elements (`option` or `input` + `label`).

That way, it won't break on options with angle brackets - like `{values: {refreshed_values: [[null, "<Script error>"]], checked_value: ""}}`.

Should be otherwise identical, except we're no longed passing the `onclick=` part of the onClickString argument (and passing it to new Function instead of creating the node with it).
[Original commit 30e5964f0f7a461aecee5f7e6a06e07466f5b7f5]