No need to replace left side tree when loading copy objects form.
Replacing of left side tree was causing tree to be replaced with a tree that's supposed to be displayed in Namespace popup on copy screen, causing screen to hang when reloading tree after save is pressed on copy objects form.
fixed a typo introduced in cf2aa79adcae7449a079ecce4518ac085f0c9a6f that was causing it to display incorrect right cell title when copying objects in automate explorer.

https://bugzilla.redhat.com/show_bug.cgi?id=1312961
[Original commit 397be35ba2d5e0cc4cd11febc0eeb009b7acdc4e]