form_div should only be replaced for certain type of catalog items.
Only replace form_div for generic type catalog items because there are no tabs on the screen, if display in catalog is set need to redraw the screen to draw the tabs and show/hide them. Replacing form_div for the types that already have tabs on screen was causing instance variables to be reset in provisioning dialogs.

https://bugzilla.redhat.com/show_bug.cgi?id=1321631
https://bugzilla.redhat.com/show_bug.cgi?id=1322834
[Original commit 85272bd43ff416b63aeffd22ba05dd08df4923f9]