JS specs - fail on parse errors
When a js spec file has a syntax error, only a small error in the browser console is shown (nothing when running under rake test:javascript), and all the tests in that particular file are skipped.

This adds another js test, that should fail whenever a file had a syntax error, so that we can at least notice these failures.