Bind sortable properties to sortable-handle class
We should not be binding the sortable properties to general elements,
such as h2, we should rather have our own class to bind the properties
to.

This is to avoid situations, where drag & drop functionality in our
UI will stop working whenever we change h2 to h3 or vice versa.

https://bugzilla.redhat.com/show_bug.cgi?id=1330580
[Original commit 3dd7844f59bd4238fc5f348d946c63289d71dd00]