Fix broken Optimize/Planning.
The wizard for selecting VMs based on different criteria was broken by
59f0c5f in #7406.

The assumption that params[:filter_typ] is sent with
params[:filter_value] is wrong. On the contrary, each of those is sent
in a separate transaction.

This causes the wizzard never to collect the values and therefor report
cannot be created.

Fix this by updating only the values that are actually sent in the
transtaction.