Fix 2 failing dialogFieldRefresh#initializeDialogSelectPicker specs
the failing test mocks miqSelectPickerEvent in beforeEach, but calls a function that passes that triggerAutoRefresh as a callback to miqSelectPickerEvent
.. and expects triggerAutoRefresh to be called

making the mock actually call the callback fixes the tests

Failures:

dialogFieldRefresh #initializeDialogSelectPicker triggers the auto refresh when the drop down changes
Expected spy triggerAutoRefresh to have been called with [ 'fieldId', 'true' ] but it was never called.

dialogFieldRefresh #initializeDialogSelectPicker triggers autorefresh with "false" when triggerAutoRefresh arg is false
Expected spy triggerAutoRefresh to have been called with [ 'fieldId', 'false' ] but it was never called.