Prevent newlines from being stripped out of RSA key upload.
The form was using the password field type in order to
cause the rsa key to be filtered out of the rails logs like
a password normally would be, but that input type was
pulling out newlines from the input. This fixes that problem,
but the file won't be filtered out of the logs.
[Original commit 526e00a5eab57bf2b45b1b9e12b2f76ac0350fdb]