Fix for never ending spinner when creating service catalog item
In case we're creating a service catalog item for a provider type,
which doesn't have any vm templates available, we'd be getting
an infinite spinner in UI.

This was caused by us relying on presence of @edit[:new][:src_vm_id]
variable, which in this particular case would be missing. This in
the end would cause a javascript error -> infinite spinner.

https://bugzilla.redhat.com/show_bug.cgi?id=1336888
[Original commit 613cb4bcf57fbd97b2eb52e947f49686e81a6929]