Split input parameters in method editor into it's own view.
Replacing whole form when adding/removing input parameters in method editor was causing to loose value in Data field on the form due to the fact that data field does not get saved in @edit[:new] until add/save button is pressed due to the large amount of data that can be saved in that field and we don't send the large amount of data up to server every time value is changed in data field on the form. To resolve the issue splitting input_div into it's own view so page replace can be done on that specific div when adding/removing Input parameters to refresh the table in method editor.

https://bugzilla.redhat.com/show_bug.cgi?id=1338007
[Original commit bd0dcb2bec3510f85a815d7e0f372822026bcb03]