Improve the validation check for Hawkular endpoins.
Currently on provider addition we only check the Hawkular endpoint presence (and status). This doesn't actually validate that the provided token can access and collect the metrics.
We should improve the validation check including also the token correctness.

This commit checks the private metric path (and not the public status) that requires valid authentication.

Issue:
https://github.com/ManageIQ/manageiq/issues/9022
[Original commit c57d07447dc8c94dd6c0b45671891db09ab64249]