Check if record supports retirement task only for VM or Template records
Made changes to code to not check for retirement task eligibility for Service & Orchestration template records. This check should only be for VmOrTemplate records. Added model methods to check if record type supports retirement. Having service record with a same id as a MiqTemple record in database was causing an issue while trying to retire the service and incorrectly going thru the code path trying to render a flash message that "Retire does not apply to selected xxxx"

https://bugzilla.redhat.com/show_bug.cgi?id=1344678
https://bugzilla.redhat.com/show_bug.cgi?id=1349989
[Original commit f4cb6ebb6aead57a323b00bd3155cf57aeb9114a]