Limit tiered rates "Per Time" to only hourly
Chargeback reports are generated by pulling hourly metrics, selecting rates and converting the selected rate to hourly and applying the selected to the hourly metric value.
This does not work very well with tiered rates that are not hourly because it will use the hourly metric to select a tier that may be daily, weekly or monthly.
This change mitigates the this by limiting to only hourly when adding or editing rate tiers.

https://bugzilla.redhat.com/show_bug.cgi?id=1337278
[Original commit 9283ea7b61811be33b8dfbac2c4bf53dad68eb1b]