No indication of failed data gathering from Hawkular endpoint
When gathering hawkular metrics fails, user has no indication in the status ui page.

This PR updates the status of the Hawkular endpoint each time metrics is collected.
It also display the last update date and time as an html title for the endpoint status, so users can know when the
status was last checked.

Bugzzila
https://bugzilla.redhat.com/show_bug.cgi?id=1344308