Introduce UI for creating VMware Cloud provider
With this patch, it is possible to create new and update existing VMware
cloud manager from the UI. Configuration form validation is similar to
other providers and a default port for the endpoint is given by the
controller.

It also updates the vmware cloud manager to allow setting the desired
API version of the backend.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>
[Original commit 4c10ada2dcf1b0e2340dfe91ba890219945eefa0]