Middleware server operations dialog: refactor http request out of controller and into stateless service.
Fixes #9462. The controller should not have any responsibilities regarding the fetching of data. This will make testing much easier since we can now mock out the service and also test it individually. The service is pure and only knows about retrieving data (and no knowledge of html forms).  When we upgrade to Angular 2 having the services separate will make for a much easier migration especially since controllers and scope don't exist in Angular 2.
[Original commit b56adb2858b32496dff552847523e0ca56942869]