miq_dynatree - make checkboxes work even when cfme_no_click
This fixes a bug introduced in #9943, where a tree with cfme_no_click would never do anything on click.

Probably the only tree affected is `protect_tree` (`Policy > Manage Policies` on a VM), without this, it's impossible to check a checkbox there.

In the original code, `onClick: function...` was created twice and this case is the one where both were created, so the second one would override the first. Changing the logic to fit that behaviour.