Specify ID nil for url_for lists
By default, rails are adding params[:id] into the url_for,
if the controller is the same as a current controller.

So if we try to show cloud_subnets in a cloud_subnets controller,
the url in a table ends up like controller/method/id/id

We need to specify that id is a nil, for these cases.
[Original commit a7ae68d323f2eb9fae287e3d6229c49616af1344]