Check the request method before using requet url as startup page
On rare events when an ajax request trigers the session timout, and loads the
login page, the startup page will be a page not found error.
This PR prevent the page not found error in this cases. In the cases of an
ajax POST request that trigers the timeout, we will replay with an unautherized replay.

Bugzilla
https://bugzilla.redhat.com/show_bug.cgi?id=1341664
[Original commit 0022001db5be7b6212a4ff927698e6c21078f56e]