Add hardware state for OpenStack Hosts
Adds an introspected and a provision_state attribute to the
Hardware model. Introspected means Ironic introspection has already
been run and has gathered the hardware details. Provisioning state
indicates whether the node is available for use or is undergoing
some sort of maintenance.

The attributes are needed to indicate where each node is at in the
Ironic workflow. Typically nodes are added an placed in manageable
state. Nodes are then introspected. And then finally placed in the
available state when ready for use.

The patch also adds the new attributes as additional fields to the
Host summary page.
[Original commit 2d0ec513f73dbe331f77b60c34074969c9c99db3]