Enable VM reconfigure disks for supported rhevm version
The VM reconfigure disks is supported only by API v4 of version
which was introduced since rhevm 4.0. The functionality will not
be available for lower levels in the 'VM reconfigure' dialog.
Also the feature requires the vm to have a store type associated
with it, so it can be used as a target storage for the new disks.

http://bugzilla.redhat.com/1373916