Host Aggregate UI phase 1
This PR includes the first phase of the Host Aggregate UI.
The goal is to achieve parity with the Availability Zone UI.
CRUD is not yet supported (as with Availabiliy zone), and K/V pair
metadata is not yet shown.

This includes most of the necessary code for C&U graphs. I've
removed the UI link to bring them up, though, since while the
graph would show, it was blank with no data in it. Getting the
CRUD bits working is a higher priority, so this PR includes basic
Host Aggregate UI, but the graphs will be made functional in a
later PR.
[Original commit d3ce5540943b4e19b4f6e35d68e4890acca35651]