rubocop issues: tree_builder_report_schedules.rb
== app/presenters/tree_builder_report_schedules.rb ==
W:  4: 25: Unused method argument - tree_name. If it's necessary, use _
or _tree_name as an argument name to indicate that it won't be used. You
can also write as tree_init_options(*) if you want the method to accept
any arguments but don't care about them.
C: 25:  5: Use the return of the conditional for variable assignment and
comparison.
[Original commit 0f8207de30a6c0a43b5c06191ebed018e8f274e3]