Skip asset_path if given absolute icon path
Calling ActionController::Base.helpers.image_path is causing significant
CPU usage for each tree node when looking up the asset locations for
non-sprockets based images.  On large trees, this accounts for a
significant CPU churn when resolving the the image path for a node.
Here is a truncation of a stackprof output as an example:

==================================
  Mode: wall(1000)
  Samples: 61211 (1.94% miss rate)
  GC: 13567 (22.16%)
==================================
     TOTAL    (pct)     SAMPLES    (pct)     FRAME
      2984   (4.9%)        2984   (4.9%)     Puma::Single#run
      1605   (2.6%)        1591   (2.6%)     block in ActiveRecord::ConnectionAdapters::PostgreSQLAdapter#exec_no_cache
      1543   (2.5%)        1543   (2.5%)     block in ActiveRecord::Result#hash_rows
>>>>  1339   (2.2%)        1339   (2.2%)     URI::RFC2396_Parser#escape
      2256   (3.7%)        1299   (2.1%)     Set#initialize

A most of the CPU churn is actually in `URI::RFC2396_Parser#escape`,
which is called by `Sprockets` under the hood to resolve asset locations
on the file system.  Specifically in `/service/explorer`, where images
can be user uploaded, the images that are returned are actually relative
paths, meaning they would not be included in the file digest of
sprockets.

If given a image starting with `/`, we basically are saying we already
know the correct location of the asset, please display that since it is
not part of the `100` directory.

Here are some quick before and after metrics for total request time on
`/service/explorer` on a ~10k node tree (these requests were made right
after an app restart):

    Before:
      request_01: 11311.114ms
      request_02:  9189.115ms
      request_03:  8837.498ms
      request_04:  8492.821ms
      request_05:  8484.918ms

    After:
      request_01: 4484.604ms
      request_02: 3645.153ms
      request_03: 3522.228ms
      request_04: 3497.319ms
      request_05: 4552.786ms

***Note***:  These metrics were taken after some other patches were
applied to speed up this page in a different way.  Your millage may very
when testing this yourself
[Original commit 3dcc57d2c22ea3f6064872742cc2b1744d4c996d]