Fix existing specs so that they use the zone called "default"
Before, we were using the zone created by FactoryGirl which
would have the name "Zone n" which would not be recognized
by the config_activated callback here
(https://github.com/ManageIQ/manageiq/blob/master/app/models/miq_server/configuration_management.rb#L33-L48)

This callback would overwrite the zone for the server using the
zone name in the settings or nil if the specified zone could not
be found. The zone was not found for these cases because the default
settings specify the zone name "default".

This lead to a failure when trying to enqueue the messages for
reloading the config when trying to find a server's zone name here
(https://github.com/ManageIQ/manageiq/blob/master/app/models/miq_server/queue_management.rb#L39)
[Original commit 4488a4518a9ff726f4e1c51d649455c9bbe0feff]