Move most of TreeNodeBuilder#build to hash
Large case statements like this require invoking the `===` method, which
in some cases can invoke Draper's `===` on presenters, which isn't very
cheap since it is coded like this:

    def ===(other)
      super || (other.respond_to?(:object) && super(other.object))
    end

The `respond_to?` call is very heavy weight, as it ends up calling
`ActiveModel::AttributeMethods#respond_to?` a bunch and dives deep in to
the model structure.  Also, it is the equivalent of an `if`/`else`,
meaning it does a comparison against the objects it won't even be using.

Since this is just syntactic sugar for saying "is this object part an
instance of this class", doing a hash lookup speeds up this process
significantly, and there is hardly a change in how the code looks.  The
biggest change is in the build method that does the lookup:

    def build
      instance_exec &BUILD_NODE_HASH[object.class]

In which the proc defined in the `BUILD_NODE_HASH` is executed within
the context of the current instance.

Decorators that might be passed in as `object` in the `#build` method
are also handled on class load.  Each key in the BUILD_NODE_HASH is
looped through to check and see if a corresponding decorator is defined
for it.  If it is, the key is added to the hash, and is mapped to the
model's node proc.
[Original commit 4ca079a8a03b1aedf0df4d1a141c45f41fb5adaa]