Fix Assumption that All Cloud Managers Support Storage Managers
This code previously assumed that a count of storage_managers could be
obtained from a cloud manager.  This is not the case when the CloudManager
did not yet have a StorageManager implemented.  Currently only Openstack
supports the StorageManager concept so this caused a failure in a spec
test for an Amazon Cloud Provider.