miqAjaxButton - make it possible to override sparkle
calling `miqAjaxButton` ends up calling `miqJqueryRequest` with `{ beforeSend: true, complete: true }` - which is equivalent to `miqSparkleOn()` before the request and `miqSparkleOff()` after.

Sometimes we do need to override this behaviour (for example, make it *not* `miqSparkleOff()` because you want to wait for a task).

This makes it possible to add a third (optional) argument to `miqAjaxButton` where you can override these (by passing `{ beforeSend: false }` for example).

Affected functions (added an optional options argument):

   * `miqService.miqAjaxButton`
   * `miqAjaxButton`
   * `miqAjaxButtonSend`
   * `miqAjax`
[Original commit 2ffad4cb1f08b663eb279b3b999c75f8f7277901]