Explorer js - reloadToolbars being empty array means don't update toolbars, not remove
Right now, when the ExplorerPresenter genereates `data.reloadToolbars` to be an empty array, the toolbar simply disappears.

Instead, it should not be touched - adding a length check to the array.

Also, the `isObject` code path exists solely for backward compatibility purposes, adding a big warning to catch these and a FIXME to remove eventually.