Use VmOrTemplate instead of MiqTemplate for lookup
TreeNodeBuilder#build has recently be switched to using a hash lookup
instead of a case.  This was to improve performance, but it did switch
the comparison that was done in the case to be assumed when doing the
lookup in `.build`, which is less forgiving.

For example, given the following object:

object = ManageIQ::Providers::Microsoft::InfraManager::Template.new

The `.base_class` of it would return `VmOrTemplate`.  This is not part
of the `TreeNodeBuilder::BUILD_NODE_HASH` keys, and neither is
`ManageIQ::Providers::Microsoft::InfraManager::Template`, so `.build`
would return `nil`.  When `.build` was a case statement, `object` would
have had `MiqTemplate::===` called against it, which would have matched
to `object`, and resolved.

`MiqTemplate` inherits from `VmOrTemplate`, and `VmOrTemplate` inherits
from `ApplicationRecord`.  Due to the functionality of `.base_class`:

http://api.rubyonrails.org/classes/ActiveRecord/Inheritance/ClassMethods.html#method-i-base_class

By switching to `"VmOrTemplate"` for the key, anything that would have
resolved to `MiqTemplate` will now resolve to `"VmOrTemplate"`, and
anything that resolves to `"VmOrTemplate"` and didn't work now will.
[Original commit 54e254ee0dc6855158191b1c6d8e428668a8e777]