Disable html5 mode for the toolbar angular module
When the current url contains a query string parameter with plus signs (like `?flash_msg=Edit+of+Cloud+Provider+"Amazon"+was+cancelled+by+the+user`), clicking on the toolbar would first mangle the url to use `%20` instead of `+`, and then clicking again would work.

This is because in html5 mode, angular needs to controll the URL.

However, there's no reason the toolbar should even be aware of the current URL, etc.

Thus, disabling html5 mode in that module.

Closes #11537
[Original commit 5c23d7fc54a8f960340d2f8989a7e4a1da7c362b]