Don't install PhantomJS on CI from BitBucket
The gemified version of Jasmine uses the gemified version of PhantomJS
which auto-installs it if it can't find your installation in ~/.phantomjs
Travis already has a version of PhantomJS installed in a different
location, so the gem will auto-install even if it's pointless.  Also,
gemified PhantomJS hardcodes install URLs from BitBucket which times out
and causes failed builds.

TLDR: Don't install auto-install PhantomJS on CI. In Travis we trust.
[Original commit 445bad97c944fbbbfbb5e2d23a4d951ec18d3e23]