StorageController#show - set x_node when called with an id
going to a specific datastore from lists that live outside of `storage_controller` - for example from `/ems_infra/123?display=storages` - works by linking to /storage/show/456

However, that path updates x_node before the trees are build and so before the right tree is selected. Forcing to :storage_tree in such a case.

https://bugzilla.redhat.com/show_bug.cgi?id=1380707
[Original commit 9291c3227fee5bbabda697e6bdb59cd30b96b119]