Add comprehensive specs for TreeNodeBuilder#build
Adds tests for (almost) every possible class the could be run through
the `.build` method from our application, and confirms the result object
from it (instead of just confirming it isn't `nil`).

This will allow for confidence in refactoring in the future to move away
from the current `hash`/`case` setup and confirm functionality.  These
specs can be moved elsewhere if such a refactoring were to take place,
but for now, this allows us to confirm existing functionality with the
current behavior.

The models added here were gathered through the following method:

1. First, the following commit was checked out to revert the changes to
TreeNodeBuilder prior to converting the large `case` in the `.build` to
a hash:

    $ git checkout ee78d9c -- app/presenters/tree_node_builder.rb

2. Then each of the keys in the `case` was run through the following
code in a `rails console` to determine which models needed to be tested
against:

    irb(main):001:0> checker_proc =  Proc.new {|klass| klass.descendants.map(&:base_class).map(&:to_s).uniq }
    => #<Proc:0x007f8febe34360@(irb):1>
    irb(main):002:0> descendants_proc = Proc.new {|klass| klass.descendants.map(&:to_s) }
    => #<Proc:0x007f8feac69748@(irb):2>
    irb(main):003:0> check = Proc.new {|klass| puts descendants_proc.call(klass).inspect; puts checker_proc.call(klass).inspect; nil }
    => #<Proc:0x007f8feaf4b0d8@(irb):3>
    irb(main):004:0> check.call(AvailabilityZone)
    ["ManageIQ::Providers::Azure::CloudManager::AvailabilityZone",
    "ManageIQ::Providers::Google::CloudManager::AvailabilityZone",
    "ManageIQ::Providers::Openstack::CloudManager::AvailabilityZone",
    "ManageIQ::Providers::Vmware::CloudManager::AvailabilityZone",
    "ManageIQ::Providers::Amazon::CloudManager::AvailabilityZone",
    "ManageIQ::Providers::Openstack::CloudManager::AvailabilityZoneNull"]
    ["AvailabilityZone"]
    => nil
    irb(main):004:0> AvailabilityZone.base_class

This allowed it to be determined the class that needed to be included in
the specs to confirm previous behavior prior to the commit.  It was then
used to validate that the new changes worked in the same fashion.

Also of note:  The "VmOrTemplate node" context in the specs was
referring to the `MiqTemplate` key from the original case statement
(changed in 54e254e for a bug fix), but was changed to reflect the
current implementation (will be updated in the future as well, since the
`Vm` key is also a subclass of `VmOrTemplate`).