Making sure that converted buttons do not get hidden by toolbar_builder methods
Buttons in 'diagnostics_servers_roles' and 'diagnostics_roles_servers'
that were converted to separate classes were still getting hide,
because in `toolbar_builder.rb` the method `build_toolbar_hide_button_ops`
returns `true` for the buttons that are not listed in the case.

By setting `false` and listing the buttons that are not converted yet
in the case switch, the converted buttons are getting displayed correctly.
[Original commit eadc3fcc6a71d064d5e2be60ccba56859e23fea3]