Move the logic around splitting requests by region into the model
This behavior is specific to VmReconfigureRequest so every call to
that model's .make_request method should handle this case.

Now `VmReconfigureRequest.make_request` will return an array of requests
if multiple were made, but just a single request object if all source
ids were in the current region to maintain backwards compatibility.
[Original commit c25d6a7da21b1d0dc16335f65cd81836fc9b38d8]