Fixed code that was gathering events for each selected category
Changed code that was looping thru records to gather timeline events for each selected category, previously it was skipping thru some of the categories with the way code was looping thru random events in Ruport dfata table. Also replaced double-quotes in the timeline event data with single-quote, new timeline widget does not like double quotes in the description.
Added spec test to verify counts of events.

https://bugzilla.redhat.com/show_bug.cgi?id=1391621