Add extra X-Angular-Request header to $http requests, don't use these to remember tab
This is to address an issue found in #12117, where Angular $http requests were remembered by remember_tab (the functionality that handles clicking on the first level menu items).

That's beacuse those are not xml_http_requests and do pretend to ask for html content, so without this we have no way to distinguish them from the ones that should be remembered.

This will not affect angular access to the API nor downloading static templates, but may affect other places if there are any that use $http to actually go somewhere else.
[Original commit 948757555078775929cfa1147299fd0368a546df]