Move user's start_url to Dashboard controller
It is only used by Dashboard controller.

If we ever need to have each page referencing start_url, we may use
javascript on client to keep the start_url for us. That way, we
calculate it only once, and we have it always available. In that case
dashboard controller will calculate the start_url only once.

Also, creating this as a mixin as we have about 3 methods tightly couple
that are good fit for separate mixin.
[Original commit b6d9b0f3ccd45b10cb07a997f25c099906c36818]