Fix Default Menu Spec Test After Removing NetApp Submenu
Previously the NetApp menu would show up for testing based on a configuration setting.
In order to allow the Roles configuration to work correctly, the empty menu used
when the NetApp configuration was disabled needed to be removed.  The test for the
NetApp item is removed via this commit. The remainder of the NetApp menu info is left
in the file for future use.
[Original commit e1c02b866e72c326989ef8f6e3a060bb0afd200a]