Fix policy x_node after create/copy/delete
In Control explorer, fixes "Containernode", "Containergroup" etc
in right side titles to be "Node", "Pod" etc (or their translations)
immediately after creating, copying or deleting a policy.

Fixes #10500, which was the last known instance of
https://bugzilla.redhat.com/show_bug.cgi?id=1359909

Modified spec to use 2-word ContainerGroup instead of Host which does
fail without this fix and works with it.
[Original commit 2dd47b004f620fcdfd18283b9870d58a395fba97]